Functional Implementation of A* (#45)
* Implementation of A*

As titled. This is the implementation of the A* algorithm instead of the Dijkstra.

* Minor test changes

As titled. These are minor adjustments for the new algorithm to work with the current tests.

* Modified weights of elevators and stairs.

As titled. Here, we are increasing the weight of the elevators and stairs to avoid a path to use the elevator when nodes are on the same floor.

* Added more tests for robust algorithm.

As titled. Added additional test cases to ensure that the algorithm we are using is robust.

* Fixed routing to call the right function

As titled.

* fix(backend): tests should run normally now

* Fixed failing tests.

As titled. Removed unnecessary tests and fixed a couple ones.

* revert(backend): use self-defined Edge type in algorithm tests

---------

Co-authored-by: felixng <ngthinh302@gmail.com>