Fixed Elevators & Fixed Graph in memory (#52)
* Fixed createGraph

As titled. Fixed createGraph function to identify elevator edges to prevent changes in elevator shafts for same floor.

* Improved pathfinding endpoint

As titled. I have improved the pathfinding endpoint and added the mapping of the floors to be able to prevent elevator changes that are unpractical.

* Elevator Penalty

As tltled. I have implemented an elevator penalty to prevent unnecessary elevator changes.

* Quick fix.

As titled. Adjusted comments.

* Added Graph Initialization

As titled. I have implemented the graph initialization at server startup.

* Pathfinding endpoint redesign

As titled. I have redesigned the pathfinding endpoint to get the graph that is already initialized.

* Removed unused mapping function. 

As titled. I removed the mapping constant that we are no longer using.

* Added Graph class

As titled. I have implemented a new class in order to be able to initialize our graph at server startup for it to be fixed in memory.

* Moved to correct directory.

* Defined Graph type.

As titled. Here, we define a Graph type just like we did in shortestPath.ts

* Fixed elevators (again)

As titled. I have fixed the elevators again introducing a new penalty system. Before, we use to remove elevators from graph when function call was for a path on the same floor. However, now, any edge leading to an elevator has a penalty weight to discourage the algorithm, as well as an elevator switch penalty so that when engaged on a path the algorithm sticks to the same elevator shaft.

* style(backend): fixed styling code format

---------

Co-authored-by: felixng <ngthinh302@gmail.com>