Template Pattern implementation.  (#100)
* New Dijkstra

As titled. We have now implemented Dijkstra algorithm.

* Dijkstra implemented on map.ts

As titled. We now handle Dijkstra on the endpoint for pathfinding.

* Dijkstra in Dropdown.

As titled. We have implemented Dijkstra in the dropdown menu.

* A* template pattern

As titled. A* now follows the template pattern.

* Dijkstra template pattern.

As titled. Dijkstra now follows the template pattern.

* Added Dijkstra.

* BasePathfindingStrategy.ts added

As titled. This is the base strategy for pathfinding to help with the template pattern.

* Redesigned map.ts.

As titled. I have redesigned map.ts to work with the template pattern.